Criacao da pasta Hooks e refactor do import
