+++ praticas/aluno-online-react/src/forms/FormLogin.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from "react-router";
-import { useAuthContext } from '../contexts/AuthContext';
+import { useAuthContext } from '../hooks/useAuthcontext';
 import InputEmail from '../../components/InputEmail';
 import InputSenha from '../../components/InputSenha';
 import BotaoSubmit from '../../components/BotaoSubmit';
