+++ praticas/aluno-online-react/src/hooks/useAuthcontext.jsx
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { AuthContext } from "../contexts/AuthContext";
+
+function useAuthContext() {
+    return useContext(AuthContext);
+}
+
+export { useAuthContext }
